rename fixtures to seed
