+++ src/index.ts
@@ -1,2 +1,2 @@
-export { seed } from './fixtures'
+export { seed } from './seed'
 export * from './types'
